// src/payload-types.ts
/* tslint:disable */
/**
 * This file was automatically generated by Payload CMS.
 * DO NOT MODIFY IT BY HAND. Instead, modify your source Payload config,
 * and re-run `payload generate:types` to regenerate this file.
 */

export interface Config {
  collections: {
    users: User;
    customers: Customer;
    notifications: Notification;
  };
  globals: {};
}
export interface User {
  id: string;
  firstName: string;
  lastName: string;
  fullname?: string;
  roles?: ('super-admin' | 'editor' | 'expert')[];
  updatedAt: string;
  createdAt: string;
  email?: string;
  resetPasswordToken?: string;
  resetPasswordExpiration?: string;
  loginAttempts?: number;
  lockUntil?: string;
  password?: string;
}
export interface Customer {
  id: string;
  firstName?: string;
  lastName?: string;
  subscription?: 'none' | 'monthly' | 'yearly';
  freeConsultations?: number;
  thisCollection?: string;
  updatedAt: string;
  createdAt: string;
  email?: string;
  resetPasswordToken?: string;
  resetPasswordExpiration?: string;
  loginAttempts?: number;
  lockUntil?: string;
  password?: string;
}
export interface Notification {
  id: string;
  title: string;
  sender?:
    | {
        value: string | User;
        relationTo: 'users';
      }
    | {
        value: string | Customer;
        relationTo: 'customers';
      };
  timestamp?: string;
  recipient:
    | (
        | {
            value: string;
            relationTo: 'users';
          }
        | {
            value: string;
            relationTo: 'customers';
          }
      )[]
    | (
        | {
            value: User;
            relationTo: 'users';
          }
        | {
            value: Customer;
            relationTo: 'customers';
          }
      )[];
  readBy?:
    | (
        | {
            value: string;
            relationTo: 'users';
          }
        | {
            value: string;
            relationTo: 'customers';
          }
      )[]
    | (
        | {
            value: User;
            relationTo: 'users';
          }
        | {
            value: Customer;
            relationTo: 'customers';
          }
      )[];
  updatedAt: string;
  createdAt: string;
}